CS-5 fairly drastic rewrite of the jq side of things.
Split apart the rules into "raw" and rules that must return an array of logical resource ids (of resources that violate/assert).

Messaging only outputs the logical resource ids in combination with an explanation of the violation.

Raw output still includes the stdout of jq.

Also lined up violation counts to be per resource across non-raw jq and custom rules.  raw jq is still just +1 per violation regardless of how many resources but i think that is ok and raw rules are the exception anyways.