#1 Add support for scanning a directory for .json and .template files and aggregating the results
also drop the notion of "fatal violation" as far as the end user is concerned.  it does stop processing for that file, but just gets logged as a "failing violation".  Given that multiple files could be processed at the same time, exiting out abruptly because one particular file has a "fatal" violation is no good - thus the stop processing flag in its stead.

update the command line argument to better reflect its a path to the json instead of an actual json string