[Finishes #165899494]
This commit introduces end-to-end tests for the cfn-nag gem, focusing on
the cfn_nag binary itself. The local workflow leverages RVM to generate
a new gemset, build and install cfn-nag locally using version 0.0.01,
installing simplecov and rspec, and then executing tests tagged with
'end_to_end'.

The remote workflow (circleci) leverages a separate job and a fresh
docker container to run the specs, and follows all of the same steps as
the local workflow. The only real difference is that circleci does not
leverage RVM, as these tests are already executing in isolation via the
new docker container.