Feature/end to end rspec tests (#199)
* [Finishes #165899494]

This commit introduces end-to-end tests for the cfn-nag gem, focusing on
the cfn_nag binary itself. The local workflow leverages RVM to generate
a new gemset, build and install cfn-nag locally using version 0.0.01,
installing simplecov and rspec, and then executing tests tagged with
'end_to_end'.

The remote workflow (circleci) leverages a separate job and a fresh
docker container to run the specs, and follows all of the same steps as
the local workflow. The only real difference is that circleci does not
leverage RVM, as these tests are already executing in isolation via the
new docker container.

* [Finishes #165899494] wrapping the rvm calls in a bit of test logic, to avoid excessive error messages in circleci and to provide slightly helpful output when running e2e tests locally WITHOUT rvm installed

*  [Finishes #165899494] Adding a bit of documentation so this is more clear

* [Finishes #165899494] this commit polishes up the language in the specs.

* [Finishes #165899494] This commit adds testing to check if cfn_nag_scan
throws an exception while scanning a directory of AWS templates.

* Fixing a typo

* Changing one spec to utilize system() rather than Open3, since STDOUT doesn't need to really be messed with. Also correcting another spec that was passing, but accidentally